refactor!: move `LengthOrPercentage` into own file
BREAKING CHANGE: Remove `LengthOrPercentage` from `length` in favor
of its own file.
