+++ src/parser/length.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-use super::{percentage::Percentage, side::CanStart, *};
+use super::*;
 use crate::tokenizer::*;
 
 #[cfg(test)]
@@ -33,14 +33,6 @@ pub enum LengthUnit {
     ViewportMaximum,    // vmax
 }
 
-/// Wraps a length or percentage together
-/// Grammar: `<length-or-percentage>
-#[derive(Debug, PartialEq, Clone, Copy)]
-pub enum LengthOrPercentage {
-    Length(Length),
-    Percentage(Percentage),
-}
-
 impl FromStr for LengthUnit {
     type Err = ();
 
@@ -65,15 +57,6 @@ impl FromStr for LengthUnit {
     }
 }
 
-impl CanStart for LengthOrPercentage {
-    fn can_start(token: &Token) -> bool {
-        matches!(
-            token,
-            Token::Number(_) | Token::Dimension(_, _) | Token::Percentage(_)
-        )
-    }
-}
-
 impl Parsable for Length {
     fn parse<I: Iterator<Item = char>>(parser: &mut Parser<I>) -> Result<Self, ParsingError> {
         match parser.tokens.next() {
@@ -89,31 +72,6 @@ impl Parsable for Length {
                 _ => Err(ParsingError::wrong_token(token_at, "dimension")),
             },
             None => Err(ParsingError::end_of_file("dimension")),
-        }
-    }
-}
-
-impl Parsable for LengthOrPercentage {
-    fn parse<I: Iterator<Item = char>>(parser: &mut Parser<I>) -> Result<Self, ParsingError> {
-        match parser.tokens.peek() {
-            Some(token_at) => match token_at.token {
-                // parse length
-                Token::Number(_) | Token::Dimension(_, _) => {
-                    Ok(LengthOrPercentage::Length(parser.parse::<Length>()?))
-                }
-
-                // parse percentage
-                Token::Percentage(_) => Ok(LengthOrPercentage::Percentage(
-                    parser.parse::<Percentage>()?,
-                )),
-
-                // neither
-                _ => Err(ParsingError::wrong_token(
-                    token_at.clone(),
-                    "length or percentage",
-                )),
-            },
-            None => Err(ParsingError::end_of_file("length or percentage")),
         }
     }
 }
